Move Auth files & Update MemoryIndex view
- Moved Files: `SignUp.js` `SignIn.js` `SignOut.js` `ChangePassword.js`:
  - Moved up one level to new folder "Authentication"
  - Updated App.js import extensions for above files

- Adjusted formatting of MemoryIndex.js (JSX)

Related to #25
