+++ src/components/Memories/MemoryIndex.js
@@ -5,7 +5,7 @@ import { memoryIndex } from '../../api/memory'
 class MemoryIndex extends Component {
   constructor () {
     super()
-    console.log('in da structor')
+    console.log('In the INDEX constructor')
     // UseFUL means use state:
     this.state = {
       memories: null
@@ -15,7 +15,7 @@ class MemoryIndex extends Component {
   componentDidMount () {
     memoryIndex(this.props.user)
       .then(res => {
-        console.log('Res is: ', res)
+        console.log('In componentDidMount, res is: ', res, '')
         this.setState({ memories: res.data.memories })
       })
       .catch(console.error)
@@ -37,8 +37,10 @@ class MemoryIndex extends Component {
             <li key={memory.id}>
               <Link to=
                 {{ pathname: `/memories/${memory.id}`, state: { id: memory.id } }}>
-                {memory.title}
+                <h4>ID={memory.id} <em>{memory.title}</em></h4>
               </Link>
+              <h6>{memory.description}</h6>
+              <p>{memory.people}</p>
             </li>
           ))}
         </ul>
